abstract sanity block logic into own component
